+++ CHANGELOG.md
@@ -1,5 +1,6 @@
-1.2.0 - DEV
+1.2.0 - Beta 8
 -----------
+
  - Correct unexpected behaviour in viewpart.php if one or more headers are not set
  - Refactor message part viewer
  - Sanitize user input